Improve note generation scripts
- Use cheaper `float`s instead of `decimal`
- Use `Update()` in `NotesGenerator` to keep track of which note is playing next, rather than generating every note at the start with coroutines

+++ unity/Assets/GGJ2026/scripts/MusicScript.cs
@@ -6,7 +6,7 @@ namespace GGJ2026
     [Serializable]
     public class MusicScript
     {
-        public decimal BeatsPerMinute { get; set; }
+        public int BeatsPerMinute { get; set; }
         public List<NoteData> Notes { get; set; } = new();
     }
 }
